Day 2 solutions
The input has to be entered after the execution of the file (just copy-paste the whole chunk of code and hit enter). Works only on the specific input because the number of rows is not provided as an input, so there is no way to cut-off the process to store the input rows after taking in all of them. I just manually counted that there are 16 rows in the question, and used it directly in the code. A simple line of code can be added if the problem would preemptively provide the number of rows in the matrix (like Hackerrank and Kattis problems usually do).