Day 10 solutions
Day 10 challenge 2 solution has a slight error in it: This is because, as the output points out, one of the hexadecimal conversions (of the number 10) is off slightly, based on the requirements of the question. The conversion remains single digit (a) even though the solution they want should be double digit (0a). So add a 0 after "49d" and before "aee" in the printed solution to get the correct answer.