Temporary solution for 13b
It's a very slipshod solution for now, but I the answer can be obtained from the code. The code will list out all time picoseconds of delay time it does not work for, in ascending order, till it finds the correct answer. For some reason, it does not print the correct answer (i.e. it just stops at the number right before the answer). So just add 1 to the last number you see. Will edit later to polish up the solution.